+++ src/spec/LoaderSpec.ts
@@ -186,11 +186,11 @@ describe('A loader', () => {
       done();
       loader.dispose();
     });
-    document.body.dispatchEvent(new KeyboardEvent('keyup', { key: 'Enter'}));
+    document.body.dispatchEvent(new KeyboardEvent('keyup', { key: 'Enter' }));
   });
 
   it('can reload without building root elements', () => {
-    const loader = new ex.Loader([,,,]);
+    const loader = new ex.Loader([, , ,]);
     loader.showPlayButton();
     loader.showPlayButton();
     loader.showPlayButton();
@@ -198,5 +198,39 @@ describe('A loader', () => {
     const buttons = document.querySelectorAll('#excalibur-play');
     expect(roots.length).toBe(1);
     expect(buttons.length).toBe(1);
+  });
+
+  /**
+   *
+   */
+  function executeMouseEvent(type: string, target: HTMLElement, button: ex.Input.NativePointerButton = null, x: number = 0, y: number = 0) {
+    const evt = new PointerEvent(type, {
+      clientX: x,
+      clientY: y,
+      button: button,
+      bubbles: true
+    });
+
+    target.dispatchEvent(evt);
+  }
+
+  it('does not propagate the start button click to pointers', (done) => {
+    const engine = new ex.Engine({ width: 1000, height: 1000 });
+    const pointerHandler = jasmine.createSpy('pointerHandler');
+    engine.input.pointers.primary.on('up', pointerHandler);
+    const loader = new Loader([new ex.Graphics.ImageSource('base/src/spec/images/GraphicsTextSpec/spritefont.png')]);
+    engine.start(loader);
+
+    setTimeout(() => {
+      const btn = (loader as any)._playButton;
+      const btnClickHandler = jasmine.createSpy('btnClickHandler');
+      btn.addEventListener('pointerup', btnClickHandler);
+      const rect = btn.getBoundingClientRect();
+      executeMouseEvent('pointerup', btn as any, ex.Input.NativePointerButton.Left, rect.x + rect.width / 2, rect.y + rect.height / 2);
+
+      expect(pointerHandler).not.toHaveBeenCalled();
+      expect(btnClickHandler).toHaveBeenCalled();
+      done();
+    }, 1000);
   });
 });